pyproject.toml
[tool.poetry]
name = "models_rj_sms"
version = "0.1.0"
description = "A short description of the project"
authors = ["Pedro Marques <pedro.marques@dados.rio>"]
package-mode = false


[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
fastapi = "^0.103.0"
uvicorn = "0.23.2"
pydantic = "^2.0.0"
crewai = "0.83.0"
crewai-tools = "^0.14.0"
vertexai = "^1.71.1"
Levenshtein = "0.26.0"
requests = "^2.32.0"
loguru = "^0.7.0"
pysqlite3-binary = "^0.5.4"

[tool.poetry.group.dev.dependencies]
black = {version = "^24.3", allow-prereleases = true}
autopep8 = "^2.0.0"
ipdb = "^0.13.0"
pylint = "^3.0.0"

[tool.poetry.group.aws]
optional = true

[tool.poetry.group.aws.dependencies]
mangum = "^0.17.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
  \.toml
  |\.sh
  |\.git
  |\.ini
  |Dockerfile
  |Jenkinfile
)/
'''
